requirements.txt
streamlit==1.20.0
openai
pdfreader
bert-extractive-summarizer
transformers
torch
